requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# web.py: 2,4
Flask == 1.1.1

# ml/utils.py: 2
Pillow == 7.0.0

# web.py: 3
Werkzeug == 0.16.0

# ml/model.py: 1,2
# ml/utils.py: 1,4
torch == 1.4.0

# ml/model.py: 3
# ml/utils.py: 3
torchvision == 0.5.0
